Moved table data
