+++ src/data/MarchMadnessTableData.ts
@@ -1,0 +1,19 @@
+type TableRowDataProps = {
+  id: string
+  label: string
+}
+
+export const TableRowData: TableRowDataProps[] = [
+  {
+    id: 'stat',
+    label: 'Stat',
+  },
+  {
+    id: 'average',
+    label: 'Average',
+  },
+  {
+    id: 'total',
+    label: 'Total',
+  },
+]
